docs/projects/project1/requirements.txt
torch
transformers
faiss-cpu
numpy
tqdm
openai
pyyaml
modelscope
